Fix definition of likelihood function and add error checking.
The log-likelihood function should have been -(sum of squared diffs),
so that the best parameter value is the maximum, rather than the minimum.

Also, add a test to screen out parameters that are likely to cause the
LLF to produce NaN or Inf values, and catch R errors in the LLF.  Both
sorts of failures are assigned a default (large, negative) return value.

Finally, in cxx-likefunc, add a parameter set that tests the parameter
screening.