Process observed data points in the MC interface in chunks.
Because we use the `nleqslv` package to solve for the budget fractions,
and because `nleqslv` treats the data points as a giant coupled system
of equations to solve, the solution time scales nonlinearly with the
number of data points.  In actuality, each data point is independent,
so breaking them up into smaller chunks causes the whole calculation
to go much faster.