mc.setup(): Return matrix of parameter guess limits.
These changes allow the Monte Carlo code to run without knowledge
of the model's parameters or their meanings.  That in turn makes it
easier to substitute any R model into the MC code.

The limits returned are used to guide the initial guesses for the
parameters, but they are not hard limits.  The MC can and in general
will stray outside of those limits as it runs.