Add data files with historical data broken out by GCAM region.
Also included is a function to write the master list out as a series
of data files.  We've used the option to get a numerical identifier
here so as to make it easier to set up a job array, but the option
exists to do name the files with the region name.