Apply symmetry condition to xi_ij values.
For the sake of expedience, we have left xi as a 2x2 matrix, but both
the app and the mc interface now pass only one cross-elasticity, which
is taken to be xi_ns.  The value of xi_sn in the parameter matrix is
ignored; instead, xi_sn is calculated in the demand function using the
symmetry constraint.