Regionalize A_s and A_n in Monte Carlo calculations
The underlying demand functions are unchanged; however, when we pass a
vector of parameters into the mc.likelihood() function we now expect
it to include a value of A_s and A_n for each region.  (We get the
list of regions from the input data).  We have the observational data
broken up by region, and we've inserted a shim function that looks up
the region for the data subset being evaluated and reconstructs the
old parameter vector (i.e., with just a single value each for A_s and
A_n) to pass to the old eval functions.