Revamp the code that generates the figures for paper 1.
Now generates the figures in a handy report using knitr.  One side
effect is that the various scripts now source one another assuming
that they are all in the same (current) directory.  This may cause
a problem using the food demand model in the MCMC code (which runs
from another directory).